% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_units.R
\name{browse_units}
\alias{browse_units}
\title{browse a page defining units of collectl reporting}
\usage{
browse_units(...)
}
\arguments{
\item{\dots}{passed to \link{browseURL}}
}
\value{
side effect is running browseURL
}
\description{
browse a page defining units of collectl reporting
}
\examples{
if (interactive()) {
browse_units()
}
}
