% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{cl_result_path}
\alias{cl_result_path}
\title{get full path to collectl report}
\usage{
cl_result_path(proc)
}
\arguments{
\item{proc}{an entity inheriting from "Rcollectl_process" S3 class}
}
\value{
character(1) path to report
}
\description{
get full path to collectl report
}
\examples{
example(cl_start)
}
