% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{cl_start}
\alias{cl_start}
\title{start collectl if possible}
\usage{
cl_start(target = tempfile())
}
\arguments{
\item{target}{character(1) path; destination of collectl report}
}
\value{
instance of \code{Rcollectl_process} with components \code{process} (a processx R6 instance) and
\code{target} (a file path where collectl results will be written)
}
\description{
start collectl if possible
}
\examples{
if (cl_exists()) {
  zz = cl_start()
  print(zz)
  print(cl_result_path(zz))
  cl_stop(zz)
  zz$process$is_alive()
}
}
