% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{cl_stop}
\alias{cl_stop}
\title{stop collectl via processx interrupt}
\usage{
cl_stop(proc)
}
\arguments{
\item{proc}{an entity inheriting from "Rcollectl_process" S3 class}
}
\value{
invisibly returns the input
}
\description{
stop collectl via processx interrupt
}
\examples{
example(cl_start)
}
