% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{print.Rcollectl_process}
\alias{print.Rcollectl_process}
\title{print method for Rcollectl process}
\usage{
\method{print}{Rcollectl_process}(x, ...)
}
\arguments{
\item{x}{an entity inheriting from "Rcollectl_process" S3 class}

\item{\dots}{not used}
}
\value{
invisibly returns the input
}
\description{
print method for Rcollectl process
}
\examples{
example(cl_start)
}
